refactor(evaluation): split evaluation package (Closes #56)
